added data persistence through local storage
