+++ public/electron.js
@@ -7,6 +7,7 @@ let win;
 function createWindow () {
   win = new BrowserWindow({width: 800, height: 600});
   win.loadURL('http://localhost:3000');
+  /*eslint-disable no-console*/
   installExtension(REACT_DEVELOPER_TOOLS)
     .then((name) => console.log('Added Extension: ', name))
     .catch((err) => console.log('An error occured: ', err ) );
