+++ src/index.js
@@ -5,6 +5,7 @@ import Teleop from './containers/teleopContainer';
 import Preform from './containers/preformContainer';
 import { Provider } from 'react-redux';
 import configureStore from './redux/config/configureStore';
+const { remote } = window.require('electron');
 const App = () => (
   <div>
     <Preform/>
@@ -12,9 +13,14 @@ const App = () => (
     <Teleop/>
   </div>
 );
-const store = configureStore();
+const store = configureStore(JSON.parse(localStorage.myAppState));
 ReactDOM.render(
   <Provider store={store}>
     <App/>
   </Provider>
   , document.getElementById('root'));
+
+remote.getCurrentWindow().on('close', () => {
+  const state = store.getState();
+  localStorage.myAppState = JSON.stringify(state);
+});
